GCC: -Wsign-compare

+++ src/game/Tactical/Enemy_Soldier_Save.cc
@@ -622,7 +622,7 @@ void NewWayOfLoadingCiviliansFromTempFile()
 			if (dp->bLife < dp->bLifeMax)
 			{
 				// Add 4 life for every hour that passes.
-				INT32 const new_life = MIN(dp->bLife + time_since_last_loaded / 15, dp->bLifeMax);
+				INT32 const new_life = MIN(dp->bLife + time_since_last_loaded / 15, (unsigned int)dp->bLifeMax);
 				dp->bLife = (INT8)new_life;
 			}
